Plots testing error vs tolerance
Replacing pre-calculated values for the constants of motion with full equations seems to have improved performance. I have also switched to circular and equatorial orbits to test deviation from x=+/-1 as well as e=0.
Up to tau=100, r and theta are stable within ~1e-12, and changing tolerance does have an effect.
eps>~1e-3 is needed for stability. The tolerance level appears to have no impact on the rate of divergence.