sets of geodesics and perturbations, plot acc and jerk
Note that integrating e>0 and x<1 takes much longer (~100x).
It still looks like the velocities may be integrating correctly, but starting with the wrong initial velocity offsets the average. However, there are oddities with some acc's and jerk's that have monotonically drifting averages.
Accelerations and jerks are currently calculated by the ODE. It may be better to calculate them analytically for each step in a for-loop after the integrator returns the positions and velocities.